curve/scalar: Move the fast scalar canonicity check here
Exposing this is generally useful.
